Don't setup a `privateFullQueryStore` as it's not possible without auth.
Engine now properly forbids the creation of a private store when there is no
criteria available to quantify what "private" is.  This means that the
`privateFullQueryStore` configuration which we previously scoped into a
separate Memcache key-namespace is no longer relevant.

We'll leave the `public:` prefix as a matter of not breaking existing
deployments, and as an exercise (and for compatibility) for those that embark
on a journey of enabling private caching.